Moved other chat settings UI related code
