+++ IronworksTranslator/MainWindow.ChatSettingsUI.cs
@@ -1,15 +1,263 @@
 ﻿using FontAwesome.WPF;
 using IronworksTranslator.Core;
+using Serilog;
 using System.Windows;
 using System.Windows.Controls;
 
 namespace IronworksTranslator
 {
     /// <summary>
-    /// 채팅 관련 설정 UI 코드들 중에 간단한 것들을 모아뒀음
+    /// Isolated functions related with chat settings UI for readability
     /// </summary>
     public partial class MainWindow : Window
     {
+        private void LoadChatSettings()
+        {
+            Log.Debug("Applying chat settings");
+
+            LoadChannelState(TellPanel, ironworksSettings.Chat.Tell);
+            LoadChannelState(SayPanel, ironworksSettings.Chat.Say);
+            LoadChannelState(YellPanel, ironworksSettings.Chat.Yell);
+            LoadChannelState(ShoutPanel, ironworksSettings.Chat.Shout);
+            LoadChannelState(PartyPanel, ironworksSettings.Chat.Party);
+            LoadChannelState(AlliancePanel, ironworksSettings.Chat.Alliance);
+
+            LoadChannelState(FreeCompanyPanel, ironworksSettings.Chat.FreeCompany);
+            LoadChannelState(NovicePanel, ironworksSettings.Chat.Novice);
+            LoadChannelState(LinkShell1Panel, ironworksSettings.Chat.LinkShell1);
+            LoadChannelState(LinkShell2Panel, ironworksSettings.Chat.LinkShell2);
+            LoadChannelState(LinkShell3Panel, ironworksSettings.Chat.LinkShell3);
+            LoadChannelState(LinkShell4Panel, ironworksSettings.Chat.LinkShell4);
+            LoadChannelState(LinkShell5Panel, ironworksSettings.Chat.LinkShell5);
+            LoadChannelState(LinkShell6Panel, ironworksSettings.Chat.LinkShell6);
+            LoadChannelState(LinkShell7Panel, ironworksSettings.Chat.LinkShell7);
+            LoadChannelState(LinkShell8Panel, ironworksSettings.Chat.LinkShell8);
+            LoadChannelState(CWLinkShell1Panel, ironworksSettings.Chat.CWLinkShell1);
+            LoadChannelState(CWLinkShell2Panel, ironworksSettings.Chat.CWLinkShell2);
+            LoadChannelState(CWLinkShell3Panel, ironworksSettings.Chat.CWLinkShell3);
+            LoadChannelState(CWLinkShell4Panel, ironworksSettings.Chat.CWLinkShell4);
+            LoadChannelState(CWLinkShell5Panel, ironworksSettings.Chat.CWLinkShell5);
+            LoadChannelState(CWLinkShell6Panel, ironworksSettings.Chat.CWLinkShell6);
+            LoadChannelState(CWLinkShell7Panel, ironworksSettings.Chat.CWLinkShell7);
+            LoadChannelState(CWLinkShell8Panel, ironworksSettings.Chat.CWLinkShell8);
+
+            Log.Debug("Chat settings applied");
+        }
+
+        private void LoadChannelState(StackPanel channelPanel, Settings.Channel channel)
+        {
+            var show = channel.Show;
+            foreach (var ui in channelPanel.Children)
+            {
+                if (ui is Button)
+                {
+                    var icon = ((Button)ui).Content as ImageAwesome;
+                    icon.Icon = show ? FontAwesomeIcon.Eye : FontAwesomeIcon.EyeSlash;
+                }
+                else if (ui is ComboBox)
+                {
+                    ((ComboBox)ui).SelectedIndex = (int)channel.MajorLanguage;
+                }
+            }
+        }
+
+        private void FieldGroupComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
+        {
+            if (ironworksSettings != null)
+            {
+                var languageIndex = ((ComboBox)sender).SelectedIndex;
+                ChangeMajorLanguage(ironworksSettings.Chat.Tell, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.Say, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.Yell, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.Shout, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.Party, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.Alliance, (ClientLanguage)languageIndex);
+
+                TellComboBox.SelectedIndex = languageIndex;
+                SayComboBox.SelectedIndex = languageIndex;
+                YellComboBox.SelectedIndex = languageIndex;
+                ShoutComboBox.SelectedIndex = languageIndex;
+                PartyComboBox.SelectedIndex = languageIndex;
+                AllianceComboBox.SelectedIndex = languageIndex;
+            }
+        }
+
+
+        private void DisplayFieldGroup(bool display, FontAwesomeIcon icon)
+        {
+            (TellShowButton.Content as ImageAwesome).Icon = icon;
+            (SayShowButton.Content as ImageAwesome).Icon = icon;
+            (YellShowButton.Content as ImageAwesome).Icon = icon;
+            (ShoutShowButton.Content as ImageAwesome).Icon = icon;
+            (PartyShowButton.Content as ImageAwesome).Icon = icon;
+            (AllianceShowButton.Content as ImageAwesome).Icon = icon;
+
+            ironworksSettings.Chat.Tell.Show = display;
+            ironworksSettings.Chat.Say.Show = display;
+            ironworksSettings.Chat.Yell.Show = display;
+            ironworksSettings.Chat.Shout.Show = display;
+            ironworksSettings.Chat.Party.Show = display;
+            ironworksSettings.Chat.Alliance.Show = display;
+        }
+
+        private void FieldGroupShowButton_Click(object sender, RoutedEventArgs e)
+        {
+            DisplayFieldGroup(true, FontAwesomeIcon.Eye);
+        }
+
+        private void FieldGroupHideButton_Click(object sender, RoutedEventArgs e)
+        {
+            DisplayFieldGroup(false, FontAwesomeIcon.EyeSlash);
+        }
+
+        private void ToggleChannelShowButton(ImageAwesome icon, Settings.Channel channel)
+        {
+            if (icon.Icon.Equals(FontAwesomeIcon.Eye))
+            {
+                icon.Icon = FontAwesomeIcon.EyeSlash;
+                channel.Show = false;
+            }
+            else
+            {
+                icon.Icon = FontAwesomeIcon.Eye;
+                channel.Show = true;
+            }
+        }
+
+        private void DisplayCommunityGroup(bool display, FontAwesomeIcon icon)
+        {
+            (FreeCompanyShowButton.Content as ImageAwesome).Icon = icon;
+            (NoviceShowButton.Content as ImageAwesome).Icon = icon;
+
+            ironworksSettings.Chat.FreeCompany.Show = display;
+            ironworksSettings.Chat.Novice.Show = display;
+        }
+
+        private void CommunityGroupComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
+        {
+            if (ironworksSettings != null)
+            {
+                var languageIndex = ((ComboBox)sender).SelectedIndex;
+                ChangeMajorLanguage(ironworksSettings.Chat.FreeCompany, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.Novice, (ClientLanguage)languageIndex);
+
+                FreeCompanyComboBox.SelectedIndex = languageIndex;
+                NoviceComboBox.SelectedIndex = languageIndex;
+            }
+        }
+
+        private void LinkShellGroupShowButton_Click(object sender, RoutedEventArgs e)
+        {
+            DisplayLinkShellGroup(true, FontAwesomeIcon.Eye);
+        }
+
+        private void LinkShellGroupHideButton_Click(object sender, RoutedEventArgs e)
+        {
+            DisplayLinkShellGroup(false, FontAwesomeIcon.EyeSlash);
+        }
+
+        private void DisplayLinkShellGroup(bool display, FontAwesomeIcon icon)
+        {
+            (LinkShell1ShowButton.Content as ImageAwesome).Icon = icon;
+            (LinkShell2ShowButton.Content as ImageAwesome).Icon = icon;
+            (LinkShell3ShowButton.Content as ImageAwesome).Icon = icon;
+            (LinkShell4ShowButton.Content as ImageAwesome).Icon = icon;
+            (LinkShell5ShowButton.Content as ImageAwesome).Icon = icon;
+            (LinkShell6ShowButton.Content as ImageAwesome).Icon = icon;
+            (LinkShell7ShowButton.Content as ImageAwesome).Icon = icon;
+            (LinkShell8ShowButton.Content as ImageAwesome).Icon = icon;
+
+            ironworksSettings.Chat.LinkShell1.Show = display;
+            ironworksSettings.Chat.LinkShell2.Show = display;
+            ironworksSettings.Chat.LinkShell3.Show = display;
+            ironworksSettings.Chat.LinkShell4.Show = display;
+            ironworksSettings.Chat.LinkShell5.Show = display;
+            ironworksSettings.Chat.LinkShell6.Show = display;
+            ironworksSettings.Chat.LinkShell7.Show = display;
+            ironworksSettings.Chat.LinkShell8.Show = display;
+        }
+
+        private void LinkShellGroupComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
+        {
+            if (ironworksSettings != null)
+            {
+                var languageIndex = ((ComboBox)sender).SelectedIndex;
+                ChangeMajorLanguage(ironworksSettings.Chat.LinkShell1, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.LinkShell2, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.LinkShell3, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.LinkShell4, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.LinkShell5, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.LinkShell6, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.LinkShell7, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.LinkShell8, (ClientLanguage)languageIndex);
+
+                LinkShell1ComboBox.SelectedIndex = languageIndex;
+                LinkShell2ComboBox.SelectedIndex = languageIndex;
+                LinkShell3ComboBox.SelectedIndex = languageIndex;
+                LinkShell4ComboBox.SelectedIndex = languageIndex;
+                LinkShell5ComboBox.SelectedIndex = languageIndex;
+                LinkShell6ComboBox.SelectedIndex = languageIndex;
+                LinkShell7ComboBox.SelectedIndex = languageIndex;
+                LinkShell8ComboBox.SelectedIndex = languageIndex;
+            }
+        }
+
+        private void CWLinkShellGroupShowButton_Click(object sender, RoutedEventArgs e)
+        {
+            DisplayCWLinkShellGroup(true, FontAwesomeIcon.Eye);
+        }
+
+        private void CWLinkShellGroupHideButton_Click(object sender, RoutedEventArgs e)
+        {
+            DisplayCWLinkShellGroup(false, FontAwesomeIcon.EyeSlash);
+        }
+
+        private void DisplayCWLinkShellGroup(bool display, FontAwesomeIcon icon)
+        {
+            (CWLinkShell1ShowButton.Content as ImageAwesome).Icon = icon;
+            (CWLinkShell2ShowButton.Content as ImageAwesome).Icon = icon;
+            (CWLinkShell3ShowButton.Content as ImageAwesome).Icon = icon;
+            (CWLinkShell4ShowButton.Content as ImageAwesome).Icon = icon;
+            (CWLinkShell5ShowButton.Content as ImageAwesome).Icon = icon;
+            (CWLinkShell6ShowButton.Content as ImageAwesome).Icon = icon;
+            (CWLinkShell7ShowButton.Content as ImageAwesome).Icon = icon;
+            (CWLinkShell8ShowButton.Content as ImageAwesome).Icon = icon;
+
+            ironworksSettings.Chat.CWLinkShell1.Show = display;
+            ironworksSettings.Chat.CWLinkShell2.Show = display;
+            ironworksSettings.Chat.CWLinkShell3.Show = display;
+            ironworksSettings.Chat.CWLinkShell4.Show = display;
+            ironworksSettings.Chat.CWLinkShell5.Show = display;
+            ironworksSettings.Chat.CWLinkShell6.Show = display;
+            ironworksSettings.Chat.CWLinkShell7.Show = display;
+            ironworksSettings.Chat.CWLinkShell8.Show = display;
+        }
+
+        private void CWLinkShellGroupComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
+        {
+            if (ironworksSettings != null)
+            {
+                var languageIndex = ((ComboBox)sender).SelectedIndex;
+                ChangeMajorLanguage(ironworksSettings.Chat.CWLinkShell1, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.CWLinkShell2, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.CWLinkShell3, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.CWLinkShell4, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.CWLinkShell5, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.CWLinkShell6, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.CWLinkShell7, (ClientLanguage)languageIndex);
+                ChangeMajorLanguage(ironworksSettings.Chat.CWLinkShell8, (ClientLanguage)languageIndex);
+
+                CWLinkShell1ComboBox.SelectedIndex = languageIndex;
+                CWLinkShell2ComboBox.SelectedIndex = languageIndex;
+                CWLinkShell3ComboBox.SelectedIndex = languageIndex;
+                CWLinkShell4ComboBox.SelectedIndex = languageIndex;
+                CWLinkShell5ComboBox.SelectedIndex = languageIndex;
+                CWLinkShell6ComboBox.SelectedIndex = languageIndex;
+                CWLinkShell7ComboBox.SelectedIndex = languageIndex;
+                CWLinkShell8ComboBox.SelectedIndex = languageIndex;
+            }
+        }
+
         private void TellComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
         {
             if (ironworksSettings != null)
